// src/components/contacts/types.ts
export interface ContactData {
  firstName: string;
  lastName: string;
  name: string;
  email: string;
  phoneNumber: string;
  jobTitle: string;
  company: string;
  industry: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
  country: string;
  status: string;
  source: string;
  lastContactedAt: string;
}